// planner-web/src/api/generated/model/createRecurringExpenseDtoFrequency.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Financial Planner API
 * Versioned API for normalized financial plans
 * OpenAPI spec version: 1.0
 */

export type CreateRecurringExpenseDtoFrequency =
  (typeof CreateRecurringExpenseDtoFrequency)[keyof typeof CreateRecurringExpenseDtoFrequency]

export const CreateRecurringExpenseDtoFrequency = {
  monthly: "monthly",
  twice_monthly: "twice_monthly",
  yearly: "yearly",
  per_payment_period: "per_payment_period",
  monthly_until_liquidated: "monthly_until_liquidated",
  custom: "custom",
} as const